Pretty straightforward, similar to implicit integrator. There's more to be gained by fusing in advance as well but I need to figure out interface considerations/composability first.

Timings are pretty similar for a single humanoid, but for 5 humanoids there is a good speedup:

before:
```
Summary for 8192 parallel rollouts

 Total JIT time: 1.26 s
 Total simulation time: 12.44 s
 Total steps per second: 658,355
 Total realtime factor: 3,291.78 x
 Total time per step: 1518.94 ns
```

after:
```
Summary for 8192 parallel rollouts

 Total JIT time: 1.28 s
 Total simulation time: 11.32 s
 Total steps per second: 723,993
 Total realtime factor: 3,619.96 x
 Total time per step: 1381.23 ns
```